+++ app/src/main/java/com/twoday/todaytrip/ui/place_map/PlaceMapAdapter.kt
@@ -9,7 +9,7 @@ import com.bumptech.glide.Glide
 import com.twoday.todaytrip.MyApplication
 import com.twoday.todaytrip.R
 import com.twoday.todaytrip.databinding.ItemPlaceMapListBinding
-import com.twoday.todaytrip.place_list_adapter.OnTourItemAddClickListener
+import com.twoday.todaytrip.ui.place_list.adapter.OnTourItemAddClickListener
 import com.twoday.todaytrip.tourData.TourItem
 import com.twoday.todaytrip.utils.TourItemPrefUtil
