+++ app/src/main/java/com/twoday/todaytrip/ui/record/RecordDetailMapAdapter.kt
@@ -9,10 +9,8 @@ import com.bumptech.glide.Glide
 import com.twoday.todaytrip.MyApplication
 import com.twoday.todaytrip.R
 import com.twoday.todaytrip.databinding.ItemPlaceMapListBinding
-import com.twoday.todaytrip.place_list_adapter.OnTourItemAddClickListener
+import com.twoday.todaytrip.ui.place_list.adapter.OnTourItemAddClickListener
 import com.twoday.todaytrip.tourData.TourItem
-import com.twoday.todaytrip.utils.RecordPrefUtil.loadRecordList
-import com.twoday.todaytrip.utils.TourItemPrefUtil
 
 class RecordDetailMapAdapter() : ListAdapter<TourItem, RecordDetailMapAdapter.Holder>(
     TourItemDiffCallback()
